Create C# CODE
Open the .NET Framework Console application in Microsoft Visual Studio, then add 6 buttons, one combo box, one textbox, and 1 timer to form 1 page. If you run your code when you add the appropriate libraries, you will see the ports on the computer in the combo box content. I created a virtual port using the VSPE application on the computer and connected it to the virtual port. In the first stage, only the "connect" button should be active on the form1 page. I did this for button 6, then when button 6 is clicked, you can select the port you chose on the combo box. It should be connected and the other buttons should be active and the 6th button should be disabled. Other buttons send numerical values via rs232 via serial port PIC18F4550. The timer1 part prints the information from RS232 to the screen and saves it to the "veri.txt" file.