MicroC Code
In the first stage, I defined the relevant pins and variables. The important part is the interrupt. The microprocessor cannot do more than one task at the same time, so it memorizes the tasks at certain time intervals, performs the task specified in the interrupt, and returns to its task.